pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["LimpidCrypto <konermann@limpidcrypto.com>"]
readme = "README.md"
packages = [
  { include = "todo_list" },
]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
flask-cors = "^4.0.0"
pyserde = "^0.16.0"
werkzeug = "^3.0.2"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
